Change_18_02_#1

+++ ESMDatabase/dbo/Tables/LAPTOP.sql
@@ -7,7 +7,7 @@
     [Storage] VARCHAR(50) NOT NULL, 
     [Graphic] VARCHAR(50) NOT NULL,  
     [Price] MONEY NOT NULL, 
-    [Discount] MONEY NULL, 
+    [Discount] FLOAT NULL, 
     [Remain] SMALLINT NOT NULL, 
     [Detail_Path] NVARCHAR(200) NULL, 
     [Image_Path] NVARCHAR(200) NULL,
