+++ ESMDatabase/dbo/Tables/PCCPU.sql
@@ -4,7 +4,7 @@
     [Name] NVARCHAR(100) NOT NULL, 
     [Socket] VARCHAR(50) NOT NULL, 
     [Price] MONEY NOT NULL, 
-    [Discount] MONEY NULL, 
+    [Discount] FLOAT NULL, 
     [Remain] SMALLINT NOT NULL, 
     [Detail_Path] NVARCHAR(200) NULL, 
     [Image_Path] NVARCHAR(200) NULL,
