+++ ESMDatabase/dbo/Tables/PCHARDDISK.sql
@@ -5,7 +5,7 @@
     [Storage] VARCHAR(50) NOT NULL, 
     [Connect] VARCHAR(20) NOT NULL,  
     [Price] MONEY NOT NULL, 
-    [Discount] MONEY NULL, 
+    [Discount] FLOAT NULL, 
     [Remain] SMALLINT NOT NULL, 
     [Detail_Path] NVARCHAR(200) NULL, 
     [Image_Path] NVARCHAR(200) NULL,
