+++ ESMDatabase/dbo/Tables/SMARTPHONE.sql
@@ -6,7 +6,7 @@
     [RAM] VARCHAR(50) NOT NULL, 
     [Storage] VARCHAR(50) NOT NULL, 
     [Price] MONEY NOT NULL, 
-    [Discount] MONEY NULL, 
+    [Discount] FLOAT NULL, 
     [Remain] SMALLINT NOT NULL, 
     [Detail_Path] NVARCHAR(200) NULL, 
     [Image_Path] NVARCHAR(200) NULL,
